Implement streaming pseudo-RPC for GetObjects. Huge reduction in memory use.
The native RPC system reads all data into memory. This is very
expensive when reading many objects or a few large objects. An RPC
GetObjects() client consumes ~2x the data size in memory and a server
is >2x worse than that. This would be a significant impediment to
scalability.

Implement a special /GetObjects HTTP handler which copies directly
between the file and the connection.