Major refactoring of lib/filesystem package to use generic inodes.
Replace the separate lists of regular files, symlinks, files and directories
with a single EntryList for directories. Replace the separate inode tables
with a single inode table of GenericInode which is an interface type. Use
type switches to yield the specific types for each inode.

This change makes the code a lot cleaner and smaller (~200 lines). It also
makes it easy to show a listing of a directory tree with all the entries
simply sorted rather than grouped by type of inode. Listings are now similar
to ls(1) output.

It increases the allocated memory ~20% and system memory ~21% for the first
scan. The sharing ratio between one scan and the next is reduced because the
directory entries are not yet shared. The increase for the 4th scan is:
allocated: 68%, system: 58%. Future work is needed on improving the sharing.

This change also increases the time taken to encode/decode a tree by ~50%.