Call the garbage collector every 1000 directory entries scanned.
This claws back most of the remaining memory bloat from the recent
refactoring, at the cost of ~7% slower scan speed (when scanning at
full speed and not rate-limiting the I/O).
It brings system memory use for the 4th scan down ~14%, and back on-par
with the code prior to the refactoring.

Tests show that increasing the garbage collection rate reduces system
memory use but further slows down the scan speed. On a system where the
I/O is being rate-limited, background garbage collection may further
decrease the system memory consumed.

It would be nice to not have to play these tricks. Experiments with
maintaining small object pools for short-lived objects did not yield
measurable benefits.