Fixes to work on ARM.
I had to make these changes to get Dominator working on Raspberry Pi.

The most significant change is the use of syscall.O_DIRECT rather than hard
coding the constant in the fsbench code. I've only tested this with linux on
arm.

The changes in walk.go fixes a build issue, the changes to the include list are
due to running govet.

I also included a fix to the README.